+++ CHANGELOG.md
@@ -9,6 +9,10 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+### Added
+
+- Added `since` argument to `Beaker.job.follow`.
+
 ## [v1.26.15](https://github.com/allenai/beaker-py/releases/tag/v1.26.15) - 2024-05-30
 
 ### Fixed